Unify Django middleware into one; Update hooks
