+++ setup.py
@@ -17,7 +17,7 @@ setup(name='instana',
                         'opentracing>=1.2.1,<1.3',
                         'basictracer>=2.2.0'],
       entry_points={'django': ['django.core.handlers.base = instana.django:hook'],
-                    'django19': ['django.core.handlers.base = instana.django19:hook'],
+                    'django19': ['django.core.handlers.base = instana.django:hook19'],
                     'flask': ['flask = instana.flaskana:hook'],
                     'runtime': ['string = instana.runtime:hook']},
       test_suite='nose.collector',
